requirements-test.txt
pytest==5.3.5
pytest-django==3.6.0
